fix: separate home.views
views ファイルを分割
