+++ home/urls.py
@@ -1,156 +1,202 @@
 from django.urls import path
 
-from home import views
+from home.views import (view_contact, view_csv, view_main, view_notice,
+                        view_notification)
 
 app_name = 'home'
 
 urlpatterns = [
-    path('', views.IndexView.as_view(), name='index'),
-    path('signup_token', views.SignUpTokenView.as_view(), name='signup_token'),
     path(
-        'signup_token_finish', views.SignUpTokenFinishView.as_view(),
+        '',
+        view_main.IndexView.as_view(),
+        name='index'
+    ),
+    path(
+        'signup_token',
+        view_main.SignUpTokenView.as_view(),
+        name='signup_token'
+    ),
+    path(
+        'signup_token_finish',
+        view_main.SignUpTokenFinishView.as_view(),
         name='signup_token_finish'
     ),
     path(
-        'signup/<token>', views.SignUpView.as_view(), name='signup'
+        'signup/<token>',
+        view_main.SignUpView.as_view(),
+        name='signup'
     ),
     path(
-        'signup_finish', views.SignUpFinishView.as_view(), name='signup_finish'
+        'signup_finish',
+        view_main.SignUpFinishView.as_view(),
+        name='signup_finish'
     ),
-    path('login', views.LoginView.as_view(), name='login'),
-    path('logout', views.LogoutView.as_view(), name='logout'),
-    path('profile', views.ProfileView.as_view(), name='profile'),
-    path('notice', views.NoticeView.as_view(), name='notice'),
-    path('notice_delete/<pk>', views.notice_delete, name='notice_delete'),
+    path(
+        'login',
+        view_main.LoginView.as_view(),
+        name='login'
+    ),
+    path(
+        'logout',
+        view_main.LogoutView.as_view(),
+        name='logout'
+    ),
+    path(
+        'profile',
+        view_main.ProfileView.as_view(),
+        name='profile'
+    ),
+    path(
+        'notice',
+        view_notice.NoticeView.as_view(),
+        name='notice'
+    ),
+    path(
+        'notice_delete/<pk>',
+        view_notice.notice_delete,
+        name='notice_delete'
+    ),
     path(
         'notice_update/<pk>',
-        views.NoticeUpdateView.as_view(), name='notice_update'
+        view_notice.NoticeUpdateView.as_view(),
+        name='notice_update'
     ),
-    path('contact', views.ContactView.as_view(), name='contact'),
-    path('contact_kind', views.ContactKindView.as_view(), name='contact_kind'),
     path(
-        'contact_list/<pk>', views.ContactListView.as_view(),
+        'contact',
+        view_contact.ContactView.as_view(),
+        name='contact'
+    ),
+    path(
+        'contact_kind',
+        view_contact.ContactKindView.as_view(),
+        name='contact_kind'
+    ),
+    path(
+        'contact_list/<pk>',
+        view_contact.ContactListView.as_view(),
         name='contact_list'
     ),
     path(
-        'contact_detail/<pk>', views.ContactDetailView.as_view(),
+        'contact_detail/<pk>',
+        view_contact.ContactDetailView.as_view(),
         name='contact_detail'
     ),
     path(
-        'notification', views.NotificationView.as_view(),
+        'notification',
+        view_notification.NotificationView.as_view(),
         {'mode': None},
         name='notification'
     ),
     path(
         'notification/reply_to_contact/<contact_pk>',
-        views.NotificationView.as_view(),
+        view_notification.NotificationView.as_view(),
         {'mode': 'reply_to_contact'},
         name='notification_reply_to_contact'
     ),
     path(
         'notification/all',
-        views.NotificationView.as_view(),
+        view_notification.NotificationView.as_view(),
         {'mode': 'all'},
         name='notification_all'
     ),
     path(
         'notification_staff_list',
-        views.NotificationStaffListView.as_view(),
+        view_notification.NotificationStaffListView.as_view(),
         name='notification_staff_list'
     ),
     path(
         'notification_staff_list/p=<page>',
-        views.NotificationStaffListView.as_view(),
+        view_notification.NotificationStaffListView.as_view(),
         name='notification_staff_list'
     ),
     path(
         'notification_staff_detail/<pk>',
-        views.NotificationStaffDetailView.as_view(),
+        view_notification.NotificationStaffDetailView.as_view(),
         name='notification_staff_detail'
     ),
     path(
         'notification_list',
-        views.NotificationListView.as_view(),
+        view_notification.NotificationListView.as_view(),
         name='notification_list'
     ),
     path(
         'notification_list/p=<page>',
-        views.NotificationListView.as_view(),
+        view_notification.NotificationListView.as_view(),
         name='notification_list'
     ),
     path(
         'notification_detail/<pk>',
-        views.NotificationDetailView.as_view(),
+        view_notification.NotificationDetailView.as_view(),
         name='notification_detail'
     ),
     path(
         'staff_list',
-        views.StaffListView.as_view(),
+        view_main.StaffListView.as_view(),
         name='staff_list'
     ),
     path(
         'download_staff_vcards/<mode>',
-        views.download_staff_vcards,
+        view_main.download_staff_vcards,
         name='download_staff_vcards'
     ),
     path(
         'csv_group',
-        views.CsvGroupView.as_view(),
+        view_csv.CsvGroupView.as_view(),
         name='csv_group'
     ),
     path(
         'csv_group_confirm',
-        views.CsvGroupConfirmView.as_view(),
+        view_csv.CsvGroupConfirmView.as_view(),
         name='csv_group_confirm'
     ),
     path(
         'csv_group_success',
-        views.CsvGroupSuccessView.as_view(),
+        view_csv.CsvGroupSuccessView.as_view(),
         name='csv_group_success'
     ),
     path(
         'csv_group_download',
-        views.csv_group_download,
+        view_csv.csv_group_download,
         name='csv_group_download'
     ),
     path(
         'csv_contact_kind',
-        views.CsvContactKindView.as_view(),
+        view_csv.CsvContactKindView.as_view(),
         name='csv_contact_kind'
     ),
     path(
         'csv_contact_kind_confirm',
-        views.CsvContactKindConfirmView.as_view(),
+        view_csv.CsvContactKindConfirmView.as_view(),
         name='csv_contact_kind_confirm'
     ),
     path(
         'csv_contact_kind_success',
-        views.CsvContactKindSuccessView.as_view(),
+        view_csv.CsvContactKindSuccessView.as_view(),
         name='csv_contact_kind_success'
     ),
     path(
         'csv_contact_kind_download',
-        views.csv_contact_kind_download,
+        view_csv.csv_contact_kind_download,
         name='csv_contact_kind_download'
     ),
     path(
         'csv_staff_register',
-        views.CsvStaffRegisterView.as_view(),
+        view_csv.CsvStaffRegisterView.as_view(),
         name='csv_staff_register'
     ),
     path(
         'csv_staff_register_confirm',
-        views.CsvStaffRegisterConfirmView.as_view(),
+        view_csv.CsvStaffRegisterConfirmView.as_view(),
         name='csv_staff_register_confirm'
     ),
     path(
         'csv_staff_register_success',
-        views.CsvStaffRegisterSuccessView.as_view(),
+        view_csv.CsvStaffRegisterSuccessView.as_view(),
         name='csv_staff_register_success'
     ),
     path(
         'csv_staff_register_download',
-        views.csv_staff_register_download,
+        view_csv.csv_staff_register_download,
         name='csv_staff_register_download'
     ),
 ]
